Fixed Issue with Raid Option:
Variables listed osias_variables.py are not replaced/updated when a new value is listed in the VM_PROFILE at runtime. 
For example, if "CEPH": "True" is in VM_PROFILE and `CEPH = False` in osias_variables.py. 
At run, it will evaluate CEPH = False instead of True as the intended result.